==> 4_pmi_analysis.Rmd <==
---
title: "PMI Analysis"
author: "Hannah Schweren"
date: "2024-03-15"
output: html_document
---

This script provides the code for the analyis of word use with Pointwise mutual information. First, a pipeline is created to turn the text data into vocabulary and then a PMI analyis is applied to this vocabulary for each subgroup of data of interest (unmatched/matched/party/age/old subgroups). The code for the PMI analysis is based on code written by Eduardo Graells-Garrido and can be found in the following repository: https://github.com/clauwag/WikipediaGenderInequality/tree/master/notebooks
 
 
### Necessary packages

```{r message=FALSE, warning=FALSE}
library(legislatoR)
library(WikipediR)
library(rvest)
library(dplyr)
library(stringr)
library(ggplot2)
library(tm)
library(MatchIt)
library(quanteda)
library(purrr)
library(dplyr)
library(tidyverse)
library(here)



```


### Load the cleaned data
```{r}
deu <- read.csv("data/clean/deu.csv")
matched_data <- read.csv(("data/clean/matched_data.csv"))
```


### Necessary functions

```{r}

#first, the data is prepared to be in the necessary shape for the PMI analyis by the following function

prepare_vocabulary <- function(data_frame, text_column) {
  #create a copy to not change the original data
  df_copy <- data_frame
  # Remove rows with missing values in the 'sex' column
  df_copy <- df_copy[complete.cases(df_copy$sex), ]
  
  # Convert the 'sex' column to numeric (male = 0, female = 1)
  df_copy$sex <- ifelse(df_copy$sex == "male", 0, 1)
  
  # Summarize the text data by sex
  summary_df <- df_copy %>%
    group_by(sex) %>%
    summarise(text = paste(!!sym(text_column), collapse = " ")) %>%
    ungroup()
  
  # Create a corpus with the of gender as identifier
  corpus_created <- corpus(summary_df, docid_field = "sex")
  
  return(corpus_created)
}

#############################################################################################################################################################

# define words that disturb the analysis, starting with gender sensitive occupation labels

gender_neutral_replacements <- c(
  "Politikerin" = "Politiker",
  "Lehrerin" = "Lehrer",
  "Sprecherin" = "Sprecher",
  "Wissenschaftlerin" = "Wissenschaftler",
  "Direktkandidatin" = "direktkandidat",
  "Staatssekretärin" = "Staatssekretär",
  "Mitarbeiterin" = "Mitarbeiter",
  "Leiterin" = "Leiter",
  "Geschäftsführerin" ="Geschäftsführer",
  "Referentin" ="referent",
  "Spitzenkandidatin" ="Spitzenkandidat",
  "Staatsministerin" = "Staatsminister",
  "Ministerin" ="Minister",
  "Rechtsanwältin" ="Rechtsanwalt",
  "Statträtin" = "Stadtrat",
  "Generalsekretärin" ="Generalsekretär",
  "Senatorin" = "Senator",
  "Richterin" ="Richter",
  "Assistentin" = "assistent",
  "Bürgermeisterin" = "Bürgermeister",
  "Präsidentin" = "Präsident",
  "Vizepräsidentin" = "Vizepräsident",
  "Obfrau" = "Obmann",
  "Kandidatin"= "Kandidat",
  "Anwältin" = "Anwalt",
  "Apothekerin"="Apotheker",
  "Betreuerin" ="Betreuer",
  "Betriebswirtin" = "Btriebswirtin",
  "Dezernentin" ="Dezernent",
  "Diplom-Volkswirtin" ="Diplom-Volkswirt",
  "Direktorin" = "Direktor",
  "Dozentin" = "Dozent",
  "Erzieherin"="Erzieher",
  "Fachärztin" = "Facharzt",
  "Familientherapeutin" ="Familientherapeut",
  "Fotografin"="Fotograf",
  "Gründerin"="Gründer",
  "Hauswirtschaftsleiterin"="Hasuwirtschaftsleiter",
  "Journalistin" = "Journalist",
  "Juristin" = "Jurist",
  "Köchin" ="Koch",
  "Korrespondentin" ="Korrespondent",
  "Krankenpflegerin" = "Krankenpfleger",
  "Kunsthistorikerin" = "Kunsthistoriker",
  "Landrätin" ="Landrat",
  "Lebensgefährtin" = "Lebensgefährt",
  "Nachfolgerin" = "Nachfolger",
  "Oberbürgermeisterin" = "oberbürgermeister",
  "Postbeamtin" = "Postbeamte",
  "Schauspielerin" = "Schauspieler",
  "Schneiderin" = "Schneider",
  "Schriftstellerin" = "Schriftsteller",
  "Sekretärin"="Sekretär",
  "Sprachlehrerin"="Sprachlehrer",
  "Stadträtin" ="Stadtrat",
  "Supervisorin"="Supervisor",
  "Unternehmerin"="Unternehmer",
  "Verkäuferin"="Verkäufer",
  "Wirtschaftsingenieurin" ="Wirtschaftsingenieur",
  "Ärztin"="Arzt",
  "Gräfin" ="graf",
  "Partnerin"="partner",
  "Bundeskanzlerin"="Bundeskanzler",
  "Betriebswirtin"="betriebswirt",
  "industriekauffrau"="Industriekaufmann",
  "Einzelhandelskauffrau"="Einzelhandelskaufmann",
  "bankkauffrau"="Bankkaufmann",
  "Autorin" = "autor",
  "Stenotypistin" = "stenotypist",
  "Pädagogin"="Padagoge",
  "Stellvertreterin" ="Stellvertreter",
  "Bundesministerin" = "bundesminister",
  "Biologin" = "Biologe",
  "Schülerin" = "Schüler",
  "Verlegerin" = "Verleger",
  "Btriebswirtin" = "Betriebswirt",
  "Staatsrätin" = "Staatsrat",
  "Fabrikarbeiterin" = "Fabrikarbeiter",
  "Steuerberaterin" = "Steuerberater",
  "Buchhalterin" = "Buchhalter",
  "Architektin" = "Architekt",
  "Justizsenatorin" = "Justizsenator",
  "Hauswirtschaftslehrerin" = "hauswirtschaftslehrer",
  "Gymnasiallehrerin" = "Gymnasiallehrer",
  "Sozialarbeiterin" ="Sozialarbeiter",
  "Studienrätin" = "Studienrat",
  "Diplom-Kauffrau" = "Diplom-Kaufmann",
  "Soldatin" = "Soldat",
  "Landwirtin" = "Landwirt",
  "Redakteurin" = "Redakteur",
  "Professorin" = "Professor",
  "Notarin" ="Notar",
  "Prokuristin" = "Prokurist",
  "Diplom-Igenieurin" = "Diplom-Ingenieur",
  "Gewerkschaftssekretärin" = "Gewerkschaftssekretär",
  "Nationalsozialistin" = "Nationalsozialist",
  "Chefredakteurin" = "Chefredakteur",
  "Bundesgeschaftsführerin" = "Bundesgeschaftsfuhrer",
  "Kommunalpolitikerin" = "Kommunalpolitiker",
  "wurde" = "", #other words to remove
  "wurden" = "",
  "jahn" = "",
  "Gustav" =""
)

# extract names
first_words <- vapply(strsplit(deu$plain_text, "\\s+"), function(x) gsub("[,;:.!?]+$", "", x[1]), character(1))
second_words <- vapply(strsplit(deu$plain_text, "\\s+"), function(x) ifelse(length(x) > 1, gsub("[,;:.!?]+$", "", x[2]), ""), character(1))

#extract months
months <- c("januar", "februar", "märz", "april", "mai", "juni", "juli", "august", "september", "oktober", "november", "dezember", "eins", "zwei", "drei", "vier", "fünf", "sechs", "sieben", "acht", "neun", "zehn", "1949–2002", "3-598-23782-0")

#extract cities
cities <- c("berlin", "heidelberg", "hamburg", "münchen", "köln", "frankfurt", 
            "stuttgart", "düsseldorf", "dortmund", "essen", "leipzig", 
            "bremen", "dresden", "hannover", "nürnberg", "duisburg", 
            "bochum", "wuppertal", "bielefeld", "bonn", "münster", 
            "karlsruhe", "mannheim", "augsburg", "wiesbaden", "gelsenkirchen", 
            "mönchengladbach", "braunschweig", "chemnitz", "kiel", "aachen", 
            "halle", "magdeburg", "freiburg im breisgau", "krefeld", 
            "lübeck", "oberhausen", "erfurt", "mainz", "rostock", "kassel", 
            "hagen", "hamm", "saarbrücken", "mülheim an der ruhr", "potsdam", 
            "ludwigshafen", "oldenburg", "leverkusen", "osnabrück", "bonn")

#extract stopwords "Stoppwörter der Deutschen Rechtssprache" (Fobbe, 2020)
custom.stopwords <- read.csv("data/raw/SW-DE-RS_v1-0-0_Datensatz.csv")
custom.stopwords <- unlist(custom.stopwords)
custom.stopwords <- unname(custom.stopwords)
custom.stopwords <- as.character(custom.stopwords)
custom.stopwords <- custom.stopwords[custom.stopwords!=""]

#combine all words that I want to remove from the vocabulary
words_to_remove <- unique(c(first_words, second_words, months, cities, stopwords("de")))

#############################################################################################################################################################

#function to further process the corpus 

process_corpus <- function(corpus, gender_neutral_replacements, words_to_remove) {
  
  # Tokenizing the corpus and removing/replacing data that disturbs the analysis
  # punctuation, capital letters, numbers etc. are removed
  tokens <- tokens(corpus, remove_punct = TRUE, remove_numbers = TRUE, remove_symbols = TRUE) %>%
    tokens_tolower() %>%
    tokens_replace(pattern = names(gender_neutral_replacements), replacement = gender_neutral_replacements, case_insensitive = TRUE) %>%
    tokens_remove(pattern = words_to_remove, padding = FALSE) %>%
    tokens_remove(pattern=custom.stopwords)%>%
    tokens_wordstem(language = "german")%>%
    tokens_remove(pattern = words_to_remove, padding = FALSE)

  
  # Creating a Document-Feature-Matrix (DFM) with the tokens
  dfmat <- dfm(tokens)
  
  # Printing the result to check
  print(dfmat)
  
  # Converting the DFM to a data frame
  corpus_df <- convert(dfmat, to = "data.frame")
  
  # Setting the first column as row names
  rownames(corpus_df) <- corpus_df[, 1]
  
  # Removing the first column from the DataFrame
  corpus_df <- corpus_df[, -1]
  
  # Keeping only words that appear in both genders by filtering out columns with zeros
  zero_count_per_column <- colSums(corpus_df == 0)
  corpus_common <- corpus_df[, zero_count_per_column == 0]
  
  # Transposing the matrix to swap rows and columns
  corpus_transposed <- t(corpus_common)
  
  # Converting the transposed matrix back to a DataFrame and adjusting columns
  corpus_transposed <- as.data.frame(corpus_transposed)
  corpus_transposed$word <- rownames(corpus_transposed)
  rownames(corpus_transposed) <- NULL
  colnames(corpus_transposed) <- c("male", "female", "word")
  
  return(corpus_transposed)
}

#############################################################################################################################################################

# the following function conducts the PMI analyis

#main analysis uses threshhold min_p as 0.0002, for sensitivity analyis adapt values accordingly
calculate_pmi_and_filter <- function(corpus_transposed, corpus_name, min_p = 0.0002) {
  # Calculate total counts
  total_male_count <- sum(corpus_transposed$male)
  total_female_count <- sum(corpus_transposed$female)
  total_count <- total_male_count + total_female_count

  # Calculate probabilities for each class
  p_c_male <- total_male_count / total_count
  p_c_female <- total_female_count / total_count

  # Add PMI calculations to the DataFrame
  pmi_df <- corpus_transposed %>%
    mutate(p_w = (male + female) / total_count,
           p_male_w = male / total_count,
           p_female_w = female / total_count,
           PMI_male = log(p_male_w / (p_w * p_c_male)) / -log(p_male_w),
           PMI_female = log(p_female_w / (p_w * p_c_female)) / -log(p_female_w))

  # Filter for words with PMI value greater than 0.02 to only get relevant results and cap at 100 words to get top 100 PMI words
  #for female words
  filtered_female <- pmi_df %>%
    filter(p_w > min_p & PMI_female > 0.02) %>%
    arrange(desc(PMI_female))%>%
    head(100)
  #and for male words
  filtered_male <- pmi_df %>%
    filter(p_w > min_p & PMI_male > 0.02) %>%
    arrange(desc(PMI_male))%>%
    head(100)


  # Save to CSV
  write.csv(filtered_female, here("data", "pmi_list", paste0(corpus_name, "_pmi_female.csv")), row.names = FALSE)
  write.csv(filtered_male, here("data", "pmi_list", paste0(corpus_name, "_pmi_male.csv")), row.names = FALSE)

  # Return a list containing both dataframes
  return(list(top_female = filtered_female, top_male = filtered_male))
}
```

# create subgroups for further analyis

In order to look into age and party groups, we need to define those groups. The following code serves this purpose
```{r}
# create three age cohorts

cohort_assignment <- function(year) {
  if(year < 1949) {
    return("cohort1")
  } else if(year >= 1949 & year <= 1969) {
    return("cohort2")
  } else {
    return("cohort3")
  }
}


# apply the function to the column birth year
cohort_vector <- sapply(matched_data$birthyear, cohort_assignment)

# split the data based on the defined cohorts
list_of_dfs <- split(matched_data, cohort_vector)

# create one dataframe per cohort to conduct the PMI analysis for each cohort
cohort1 <- list_of_dfs$cohort1
cohort2 <- list_of_dfs$cohort2
cohort3 <- list_of_dfs$cohort3

```


```{r}

# Split up different parties in respective dataframes

parties <- c("FDP", "AfD", "BÜNDNIS 90/DIE GRÜNEN", "CDU", "CSU", "SPD", "DIE LINKE")

# list to save the individual party dataframes
party_dfs <- list()

# Loop over the party names
for(party in parties) {
  if(party == "CDU" || party == "CSU") {
    # Combine data for CDU and CSU
    if(!"CDU_CSU" %in% names(party_dfs)) {
      party_dfs[["CDU_CSU"]] <- matched_data[matched_data$party %in% c("CDU", "CSU"), ]
    }
  } else {
    # Create a DataFrame for all other parties
    party_dfs[[party]] <- matched_data[matched_data$party == party, ]
  }
}

# get separate dataframes for each political party
cdu_csu_df <- party_dfs[["CDU_CSU"]]
greens_df <- party_dfs[["BÜNDNIS 90/DIE GRÜNEN"]]
left_df <- party_dfs[["DIE LINKE"]]
fdp_df <- party_dfs[["FDP"]]
spd_df <- party_dfs[["SPD"]]
afd_df <- party_dfs[["AfD"]]

```



### Apply pipeline to all relevant groups

In the following code, the functions to output the PMI lists are applied to all relevant groups
```{r}


# #whole sample, life section########################################################################################
# # Step 1: 
corpus <- prepare_vocabulary(deu, "extracted_text")

# Step 2:
corpus_transposed <- process_corpus(corpus, gender_neutral_replacements, words_to_remove)

# Step 3:
results <- calculate_pmi_and_filter(corpus_transposed, "deu_life")


# #whole sample career/party section########################################################################################
# # Step 1: 
# corpus <- prepare_vocabulary(deu, "career_text")
# 
# # Step 2: 
# corpus_transposed <- process_corpus(corpus, gender_neutral_replacements, words_to_remove)
# 
# # Step 3: 
# results <- calculate_pmi_and_filter(corpus_transposed, "deu_career")
# 
# #matched data life section########################################################################################
# # Step 1: 
# corpus <- prepare_vocabulary(matched_data, "extracted_text")
# 
# # Step 2: 
# corpus_transposed <- process_corpus(corpus, gender_neutral_replacements, words_to_remove)
# 
# # Step 3: 
# results <- calculate_pmi_and_filter(corpus_transposed, "matched_life")
# 
# #matched data career section########################################################################################
# # Step 1: 
# corpus <- prepare_vocabulary(matched_data, "career_text")
# 
# # Step 2: 
# corpus_transposed <- process_corpus(corpus, gender_neutral_replacements, words_to_remove)
# 
# # Step 3: 
# results <- calculate_pmi_and_filter(corpus_transposed, "matched_career")
# 
# ##################################################################################################################################

####################old vs. new data 

# # #matched sample, old life section########################################################################################
# # # Step 1:
# corpus <- prepare_vocabulary(matched_data, "extracted_text_old")
# 
# # Step 2:
# corpus_transposed <- process_corpus(corpus, gender_neutral_replacements, words_to_remove)
# 
# # Step 3:
# results <- calculate_pmi_and_filter(corpus_transposed, "matched_old_life")
# 
# # #matched sample, new life section########################################################################################
# # Step 1:
# corpus <- prepare_vocabulary(matched_data, "extracted_text_new")
# 
# # Step 2:
# corpus_transposed <- process_corpus(corpus, gender_neutral_replacements, words_to_remove)
# 
# # Step 3:
# results <- calculate_pmi_and_filter(corpus_transposed, "matched_new_life")
# 
# 
# # #matched sample, old career section########################################################################################
# # # Step 1:
# corpus <- prepare_vocabulary(matched_data, "career_text_old")
# 
# # Step 2:
# corpus_transposed <- process_corpus(corpus, gender_neutral_replacements, words_to_remove)
# 
# # Step 3:
# results <- calculate_pmi_and_filter(corpus_transposed, "matched_old_career")
# 
# # #matched sample, new career section########################################################################################
# # Step 1:
# corpus <- prepare_vocabulary(matched_data, "career_text_new")
# 
# # Step 2:
# corpus_transposed <- process_corpus(corpus, gender_neutral_replacements, words_to_remove)
# 
# # Step 3:
# results <- calculate_pmi_and_filter(corpus_transposed, "matched_new_career")

####################################################################################################################
# 
# #cohort 1 career section########################################################################################
# # Step 1: 
# corpus <- prepare_vocabulary(first_cohort, "career_text")
# 
# # Step 2: 
# corpus_transposed <- process_corpus(corpus, gender_neutral_replacements, words_to_remove)
# 
# # Step 3: 
# results <- calculate_pmi_and_filter(corpus_transposed, "1_cohort_career")
# 
# 
# #cohort 2 career section########################################################################################
# # Step 1: 
# corpus <- prepare_vocabulary(second_cohort, "career_text")
# 
# # Step 2: 
# corpus_transposed <- process_corpus(corpus, gender_neutral_replacements, words_to_remove)
# 
# # Step 3: 
# results <- calculate_pmi_and_filter(corpus_transposed, "2_cohort_career")
# 
# 
# #cohort 3 career section########################################################################################
# # Step 1: 
# corpus <- prepare_vocabulary(third_cohort, "career_text")
# 
# # Step 2: 
# corpus_transposed <- process_corpus(corpus, gender_neutral_replacements, words_to_remove)
# 
# # Step 3: 
# results <- calculate_pmi_and_filter(corpus_transposed, "3_cohort_career")
# 
# 
# #cohort 4 career section########################################################################################
# # Step 1: 
# corpus <- prepare_vocabulary(fourth_cohort, "career_text")
# 
# # Step 2: 
# corpus_transposed <- process_corpus(corpus, gender_neutral_replacements, words_to_remove)
# 
# # Step 3: 
# results <- calculate_pmi_and_filter(corpus_transposed, "4_cohort_career")
# 
# 
# #cohort 5 career section########################################################################################
# # Step 1: 
# corpus <- prepare_vocabulary(fifth_cohort, "career_text")
# 
# # Step 2: 
# corpus_transposed <- process_corpus(corpus, gender_neutral_replacements, words_to_remove)
# 
# # Step 3: 
# results <- calculate_pmi_and_filter(corpus_transposed, "5_cohort_career")
# 
# #cohort 1 life section########################################################################################
# # Step 1: 
# corpus <- prepare_vocabulary(first_cohort, "extracted_text")
# 
# # Step 2: 
# corpus_transposed <- process_corpus(corpus, gender_neutral_replacements, words_to_remove)
# 
# # Step 3: 
# results <- calculate_pmi_and_filter(corpus_transposed, "1_cohort_life")
# 
# 
# #cohort 2 life section########################################################################################
# # Step 1: 
# corpus <- prepare_vocabulary(second_cohort, "extracted_text")
# 
# # Step 2: 
# corpus_transposed <- process_corpus(corpus, gender_neutral_replacements, words_to_remove)
# 
# # Step 3: 
# results <- calculate_pmi_and_filter(corpus_transposed, "2_cohort_life")
# 
# 
# #cohort 3 life section########################################################################################
# # Step 1: 
# corpus <- prepare_vocabulary(third_cohort, "extracted_text")
# 
# # Step 2: 
# corpus_transposed <- process_corpus(corpus, gender_neutral_replacements, words_to_remove)
# 
# # Step 3: 
# results <- calculate_pmi_and_filter(corpus_transposed, "3_cohort_life")
# 
# 
# #cohort 4 life section########################################################################################
# # Step 1: 
# corpus <- prepare_vocabulary(fourth_cohort, "extracted_text")
# 
# # Step 2: 
# corpus_transposed <- process_corpus(corpus, gender_neutral_replacements, words_to_remove)
# 
# # Step 3: 
# results <- calculate_pmi_and_filter(corpus_transposed, "4_cohort_life")
# 
# 
# #cohort 5 life section########################################################################################
# # Step 1: 
# corpus <- prepare_vocabulary(fifth_cohort, "extracted_text")
# 
# # Step 2: 
# corpus_transposed <- process_corpus(corpus, gender_neutral_replacements, words_to_remove)
# 
# # Step 3: 
# results <- calculate_pmi_and_filter(corpus_transposed, "5_cohort_life")
# 
# 
# ##################################################################################################################################

#cohorts - 3 groups
#cohort 3 life section########################################################################################
# # Step 1:
# corpus <- prepare_vocabulary(cohort3, "extracted_text")
# 
# # Step 2:
# corpus_transposed <- process_corpus(corpus, gender_neutral_replacements, words_to_remove)
# 
# # Step 3:
# results <- calculate_pmi_and_filter(corpus_transposed, "third_cohort_life")
# 
# 
# #cohort 2 life section########################################################################################
# # Step 1:
# corpus <- prepare_vocabulary(cohort2, "extracted_text")
# 
# # Step 2:
# corpus_transposed <- process_corpus(corpus, gender_neutral_replacements, words_to_remove)
# 
# # Step 3:
# results <- calculate_pmi_and_filter(corpus_transposed, "second_cohort_life")
# 
# #cohort 1 life section########################################################################################
# # Step 1:
# corpus <- prepare_vocabulary(cohort1, "extracted_text")
# 
# # Step 2:
# corpus_transposed <- process_corpus(corpus, gender_neutral_replacements, words_to_remove)
# 
# # Step 3:
# results <- calculate_pmi_and_filter(corpus_transposed, "first_cohort_life")
# 
# 
# #cohort 3 career section########################################################################################
# # Step 1:
# corpus <- prepare_vocabulary(cohort3, "career_text")
# 
# # Step 2:
# corpus_transposed <- process_corpus(corpus, gender_neutral_replacements, words_to_remove)
# 
# # Step 3:
# results <- calculate_pmi_and_filter(corpus_transposed, "third_cohort_career")
# 
# #cohort 2 career section########################################################################################
# # Step 1:
# corpus <- prepare_vocabulary(cohort2, "career_text")
# 
# # Step 2:
# corpus_transposed <- process_corpus(corpus, gender_neutral_replacements, words_to_remove)
# 
# # Step 3:
# results <- calculate_pmi_and_filter(corpus_transposed, "second_cohort_career")
# 
# #cohort 1 career section########################################################################################
# # Step 1:
# corpus <- prepare_vocabulary(cohort1, "career_text")
# 
# # Step 2:
# corpus_transposed <- process_corpus(corpus, gender_neutral_replacements, words_to_remove)
# 
# # Step 3:
# results <- calculate_pmi_and_filter(corpus_transposed, "first_cohort_career")
# 
# # 
# #afd life section########################################################################################
# # Step 1: 
# corpus <- prepare_vocabulary(afd_df, "extracted_text")
# 
# # Step 2: 
# corpus_transposed <- process_corpus(corpus, gender_neutral_replacements, words_to_remove)
# 
# # Step 3: 
# results <- calculate_pmi_and_filter(corpus_transposed, "afd_life")
# 
# #cdu/csu life section########################################################################################
# # Step 1: 
# corpus <- prepare_vocabulary(cdu_csu_df, "extracted_text")
# 
# # Step 2: 
# corpus_transposed <- process_corpus(corpus, gender_neutral_replacements, words_to_remove)
# 
# # Step 3: 
# results <- calculate_pmi_and_filter(corpus_transposed, "cdu_csu_life")
# 
# #greens life section########################################################################################
# # Step 1: 
# corpus <- prepare_vocabulary(greens_df, "extracted_text")
# 
# # Step 2: 
# corpus_transposed <- process_corpus(corpus, gender_neutral_replacements, words_to_remove)
# 
# # Step 3: 
# results <- calculate_pmi_and_filter(corpus_transposed, "greens_life")
# 
# #left life section########################################################################################
# # Step 1: 
# corpus <- prepare_vocabulary(left_df, "extracted_text")
# 
# # Step 2: 
# corpus_transposed <- process_corpus(corpus, gender_neutral_replacements, words_to_remove)
# 
# # Step 3: 
# results <- calculate_pmi_and_filter(corpus_transposed, "left_life")
# 
# #fdp life section########################################################################################
# # Step 1: 
# corpus <- prepare_vocabulary(fdp_df, "extracted_text")
# 
# # Step 2: 
# corpus_transposed <- process_corpus(corpus, gender_neutral_replacements, words_to_remove)
# 
# # Step 3: 
# results <- calculate_pmi_and_filter(corpus_transposed, "fdp_life")
# 
# #spd life section########################################################################################
# # Step 1: 
# corpus <- prepare_vocabulary(spd_df, "extracted_text")
# 
# # Step 2: 
# corpus_transposed <- process_corpus(corpus, gender_neutral_replacements, words_to_remove)
# 
# # Step 3: 
# results <- calculate_pmi_and_filter(corpus_transposed, "spd_life")
# 
# #afd career section########################################################################################
# # Step 1: 
# corpus <- prepare_vocabulary(afd_df, "career_text")
# 
# # Step 2: 
# corpus_transposed <- process_corpus(corpus, gender_neutral_replacements, words_to_remove)
# 
# # Step 3: 
# results <- calculate_pmi_and_filter(corpus_transposed, "afd_career")
# 
# #cdu/csu career section########################################################################################
# # Step 1: 
# corpus <- prepare_vocabulary(cdu_csu_df, "career_text")
# 
# # Step 2: 
# corpus_transposed <- process_corpus(corpus, gender_neutral_replacements, words_to_remove)
# 
# # Step 3: 
# results <- calculate_pmi_and_filter(corpus_transposed, "cdu_csu_career")
# 
# #greens career section########################################################################################
# # Step 1: 
# corpus <- prepare_vocabulary(greens_df, "career_text")
# 
# # Step 2: 
# corpus_transposed <- process_corpus(corpus, gender_neutral_replacements, words_to_remove)
# 
# # Step 3: 
# results <- calculate_pmi_and_filter(corpus_transposed, "greens_career")
# 
# #left career section########################################################################################
# # Step 1: 
# corpus <- prepare_vocabulary(left_df, "career_text")
# 
# # Step 2: 
# corpus_transposed <- process_corpus(corpus, gender_neutral_replacements, words_to_remove)
# 
# # Step 3: 
# results <- calculate_pmi_and_filter(corpus_transposed, "left_career")
# 
# #fdp career section########################################################################################
# # Step 1: 
# corpus <- prepare_vocabulary(fdp_df, "career_text")
# 
# # Step 2: 
# corpus_transposed <- process_corpus(corpus, gender_neutral_replacements, words_to_remove)
# 
# # Step 3: 
# results <- calculate_pmi_and_filter(corpus_transposed, "fdp_career")
# 
# #spd career section########################################################################################
# # Step 1: 
# corpus <- prepare_vocabulary(spd_df, "career_text")
# 
# # Step 2: 
# corpus_transposed <- process_corpus(corpus, gender_neutral_replacements, words_to_remove)
# 
# # Step 3: 
# results <- calculate_pmi_and_filter(corpus_transposed, "spd_career")




```

### Sensitivity Analysis for different thresholds
```{r}

# ######## IMOPRTANT: Set min_p to the respective threshhold to be tested #########################################
# 
# # #whole sample, life section threshold 0.001 ########################################################################################
# # Step 1:
# corpus <- prepare_vocabulary(deu, "extracted_text")
# 
# # Step 2:
# corpus_transposed <- process_corpus(corpus, gender_neutral_replacements, words_to_remove)
# 
# # Step 3:
# results <- calculate_pmi_and_filter(corpus_transposed, "S_deu_life")
# 
# 
# # #whole sample career/party section########################################################################################
# # # Step 1:
# corpus <- prepare_vocabulary(deu, "career_text")
# 
# # Step 2:
# corpus_transposed <- process_corpus(corpus, gender_neutral_replacements, words_to_remove)
# 
# # Step 3:
# results <- calculate_pmi_and_filter(corpus_transposed, "S_deu_career")
# 
# ##########################################################################################################################
# 
# # #whole sample, life section threshold 0.0007########################################################################################
# # Step 1:
# corpus <- prepare_vocabulary(deu, "extracted_text")
# 
# # Step 2:
# corpus_transposed <- process_corpus(corpus, gender_neutral_replacements, words_to_remove)
# 
# # Step 3:
# results <- calculate_pmi_and_filter(corpus_transposed, "S2_deu_life")
# 
# 
# # #whole sample career/party section########################################################################################
# # # Step 1:
# corpus <- prepare_vocabulary(deu, "career_text")
# 
# # Step 2:
# corpus_transposed <- process_corpus(corpus, gender_neutral_replacements, words_to_remove)
# 
# # Step 3:
# results <- calculate_pmi_and_filter(corpus_transposed, "S2_deu_career")
# 
# ##########################################################################################################################
# 
# # #whole sample career/party section threshold 0.0004 ########################################################################################
# # # Step 1:
# 
# corpus <- prepare_vocabulary(deu, "extracted_text")
# 
# # Step 2:
# corpus_transposed <- process_corpus(corpus, gender_neutral_replacements, words_to_remove)
# 
# # Step 3:
# results <- calculate_pmi_and_filter(corpus_transposed, "S3_deu_life")
# 
# 
# # #whole sample career/party section Threshold 0.0003########################################################################################
# # # Step 1:
# corpus <- prepare_vocabulary(deu, "career_text")
# 
# # Step 2:
# corpus_transposed <- process_corpus(corpus, gender_neutral_replacements, words_to_remove)
# 
# # Step 3:
# results <- calculate_pmi_and_filter(corpus_transposed, "S4_deu_career")
# 
# 
# 
# corpus <- prepare_vocabulary(deu, "extracted_text")
# 
# # Step 2:
# 
# # Step 3:
# results <- calculate_pmi_and_filter(corpus_transposed, "S4_deu_life")
# 
# 
# # #whole sample career/party section########################################################################################
# # # Step 1:
# corpus <- prepare_vocabulary(deu, "career_text")
# 
# # Step 2:
# corpus_transposed <- process_corpus(corpus, gender_neutral_replacements, words_to_remove)
# 
# # Step 3:
# results <- calculate_pmi_and_filter(corpus_transposed, "S4_deu_career")
```
