==> 5_pmi_visualisierung.Rmd <==
---
title: "PMI figures and tables"
author: "Hannah Schweren"
date: "2024-03-27"
output: html_document
---
Script to create figures and tables to present the results of the Pointwise Mutual Information analysis. The results can be visualised as bar plots or as tables, the latter is more useful for the bigger subgroups of interest like Party comparisons.

### Necessary packages

```{r message=FALSE, warning=FALSE}
library(ggplot2)
library(dplyr)
library(purrr)
library(here)
library(knitr)
library(kableExtra)
library(dplyr)
library(tidyr)
```

### load the annotaded data
the PMI lists were created in the previous script and annotated in excel. Now we can load these annotated lists to observe the share of categories (Family, Gender, Relatioonship and Other) for all subgroups of interest

```{r}

#matched and unmatched data

deu_life_female_cat <- read.csv(here("data", "annotated_data", "deu_life_female_cat.csv"), sep = ";")
deu_life_male_cat <- read.csv(here("data", "annotated_data", "deu_life_male_cat.csv"), sep = ";")

matched_life_female_cat <- read.csv(here("data", "annotated_data", "matched_life_female_cat.csv"), sep = ";")
matched_life_male_cat <- read.csv(here("data", "annotated_data", "matched_life_male_cat.csv"), sep = ";")

deu_career_female_cat <- read.csv(here("data", "annotated_data", "deu_career_female_cat.csv"), sep = ";")
deu_career_male_cat <- read.csv(here("data", "annotated_data", "deu_career_male_cat.csv"), sep = ";")

matched_career_female_cat <- read.csv(here("data", "annotated_data", "matched_career_female_cat.csv"), sep = ";")
matched_career_male_cat <- read.csv(here("data", "annotated_data", "matched_career_male_cat.csv"), sep = ";")


#sensitivitiy analyis

S_deu_life_female_cat <- read.csv(here("data", "annotated_data", "S_deu_life_female_cat.csv"), sep = ";")
S_deu_life_male_cat <- read.csv(here("data", "annotated_data", "S_deu_life_male_cat.csv"), sep = ";")

S_deu_career_female_cat <- read.csv(here("data", "annotated_data", "S_deu_career_female_cat.csv"), sep = ";")
S_deu_career_male_cat <- read.csv(here("data", "annotated_data", "S_deu_career_male_cat.csv"), sep = ";")

S2_deu_life_female_cat <- read.csv(here("data", "annotated_data", "S2_deu_life_female_cat.csv"), sep = ";")
S2_deu_life_male_cat <- read.csv(here("data", "annotated_data", "S2_deu_life_male_cat.csv"), sep = ";")

S2_deu_career_female_cat <- read.csv(here("data", "annotated_data", "S2_deu_career_female_cat.csv"), sep = ";")
S2_deu_career_male_cat <- read.csv(here("data", "annotated_data", "S2_deu_career_male_cat.csv"), sep = ";")

S3_deu_life_female_cat <- read.csv(here("data", "annotated_data", "S3_deu_life_female_cat.csv"), sep = ";")
S3_deu_life_male_cat <- read.csv(here("data", "annotated_data", "S3_deu_life_male_cat.csv"), sep = ";")

S3_deu_career_female_cat <- read.csv(here("data", "annotated_data", "S3_deu_career_female_cat.csv"), sep = ";")
S3_deu_career_male_cat <- read.csv(here("data", "annotated_data", "S3_deu_career_male_cat.csv"), sep = ";")

S4_deu_life_female_cat <- read.csv(here("data", "annotated_data", "S4_deu_life_female_cat.csv"), sep = ";")
S4_deu_life_male_cat <- read.csv(here("data", "annotated_data", "S4_deu_life_male_cat.csv"), sep = ";")

S4_deu_career_female_cat <- read.csv(here("data", "annotated_data", "S4_deu_career_female_cat.csv"), sep = ";")
S4_deu_career_male_cat <- read.csv(here("data", "annotated_data", "S4_deu_career_male_cat.csv"), sep = ";")


# parties

afd_life_female_cat <- read.csv(here("data", "annotated_data", "afd_life_female_cat.csv"), sep = ";")
afd_life_male_cat <- read.csv(here("data", "annotated_data", "afd_life_male_cat.csv"), sep = ";")

afd_career_female_cat <- read.csv(here("data", "annotated_data", "afd_career_female_cat.csv"), sep = ";")
afd_career_male_cat <- read.csv(here("data", "annotated_data", "afd_career_male_cat.csv"), sep = ";")

cdu_csu_life_female_cat <- read.csv(here("data", "annotated_data", "cdu_csu_life_female_cat.csv"), sep = ";")
cdu_csu_life_male_cat <- read.csv(here("data", "annotated_data", "cdu_csu_life_male_cat.csv"), sep = ";")

cdu_csu_career_female_cat <- read.csv(here("data", "annotated_data", "cdu_csu_career_female_cat.csv"), sep = ";")
cdu_csu_career_male_cat <- read.csv(here("data", "annotated_data", "cdu_csu_career_male_cat.csv"), sep = ";")

spd_life_female_cat <- read.csv(here("data", "annotated_data", "spd_life_female_cat.csv"), sep = ";")
spd_life_male_cat <- read.csv(here("data", "annotated_data", "spd_life_male_cat.csv"), sep = ";")

spd_career_female_cat <- read.csv(here("data", "annotated_data", "spd_career_female_cat.csv"), sep = ";")
spd_career_male_cat <- read.csv(here("data", "annotated_data", "spd_career_male_cat.csv"), sep = ";")

fdp_life_female_cat <- read.csv(here("data", "annotated_data", "fdp_life_female_cat.csv"), sep = ";")
fdp_life_male_cat <- read.csv(here("data", "annotated_data", "fdp_life_male_cat.csv"), sep = ";")

fdp_career_female_cat <- read.csv(here("data", "annotated_data", "fdp_career_female_cat.csv"), sep = ";")
fdp_career_male_cat <- read.csv(here("data", "annotated_data", "fdp_career_male_cat.csv"), sep = ";")

greens_life_female_cat <- read.csv(here("data", "annotated_data", "greens_life_female_cat.csv"), sep = ";")
greens_life_male_cat <- read.csv(here("data", "annotated_data", "greens_life_male_cat.csv"), sep = ";")

greens_career_female_cat <- read.csv(here("data", "annotated_data", "greens_career_female_cat.csv"), sep = ";")
greens_career_male_cat <- read.csv(here("data", "annotated_data", "greens_career_male_cat.csv"), sep = ";")

left_life_female_cat <- read.csv(here("data", "annotated_data", "left_life_female_cat.csv"), sep = ";")
left_life_male_cat <- read.csv(here("data", "annotated_data", "left_life_male_cat.csv"), sep = ";")

left_career_female_cat <- read.csv(here("data", "annotated_data", "left_career_female_cat.csv"), sep = ";")
left_career_male_cat <- read.csv(here("data", "annotated_data", "left_career_male_cat.csv"), sep = ";")


# old and new data

matched_old_life_female_cat <- read.csv(here("data", "annotated_data", "matched_old_life_female_cat.csv"), sep = ";")
matched_old_life_male_cat <- read.csv(here("data", "annotated_data", "left_life_male_cat.csv"), sep = ";")

matched_new_life_female_cat <- read.csv(here("data", "annotated_data", "matched_new_life_female_cat.csv"), sep = ";")
matched_new_life_male_cat <- read.csv(here("data", "annotated_data", "matched_new_life_male_cat.csv"), sep = ";")

matched_old_career_female_cat <- read.csv(here("data", "annotated_data", "matched_old_career_female_cat.csv"), sep = ";")
matched_old_career_male_cat <- read.csv(here("data", "annotated_data", "left_career_male_cat.csv"), sep = ";")

matched_new_career_female_cat <- read.csv(here("data", "annotated_data", "matched_new_life_female_cat.csv"), sep = ";")
matched_new_career_male_cat <- read.csv(here("data", "annotated_data", "matched_new_life_male_cat.csv"), sep = ";")


# age cohorts

first_cohort_life_female_cat <- read.csv(here("data", "annotated_data", "first_cohort_life_female_cat.csv"), sep = ";")
first_cohort_life_male_cat <- read.csv(here("data", "annotated_data", "first_cohort_life_male_cat.csv"), sep = ";")

first_cohort_career_female_cat <- read.csv(here("data", "annotated_data", "first_cohort_career_female_cat.csv"), sep = ";")
first_cohort_career_male_cat <- read.csv(here("data", "annotated_data", "first_cohort_career_male_cat.csv"), sep = ";")

second_cohort_life_female_cat <- read.csv(here("data", "annotated_data", "second_cohort_life_female_cat.csv"), sep = ";")
second_cohort_life_male_cat <- read.csv(here("data", "annotated_data", "second_cohort_life_male_cat.csv"), sep = ";")

second_cohort_career_female_cat <- read.csv(here("data", "annotated_data", "second_cohort_career_female_cat.csv"), sep = ";")
second_cohort_career_male_cat <- read.csv(here("data", "annotated_data", "second_cohort_career_male_cat.csv"), sep = ";")

third_cohort_life_female_cat <- read.csv(here("data", "annotated_data", "third_cohort_life_female_cat.csv"), sep = ";")
third_cohort_life_male_cat <- read.csv(here("data", "annotated_data", "third_cohort_life_male_cat.csv"), sep = ";")

third_cohort_career_female_cat <- read.csv(here("data", "annotated_data", "third_cohort_career_female_cat.csv"), sep = ";")
third_cohort_career_male_cat <- read.csv(here("data", "annotated_data", "third_cohort_career_male_cat.csv"), sep = ";")


```

### Necesssary functions

```{r}

# function to create a barplot of the categories assigned

pmi_plot <- function(df_female, df_male) {
  # combine male and female annotated data and add column indicating the gender 
  df_female$Gender <- 'female'
  df_male$Gender <- 'male'
  combined_df <- rbind(df_female, df_male)
  
  # calculate share of each category
  category_proportions <- combined_df %>%
    group_by(Gender, category) %>%
    summarise(count = n(), .groups = 'drop') %>%
    group_by(Gender) %>%
    mutate(proportion = count / sum(count))
  
  # conduct fishers's test in order to include the P-Value to indicate significance of the results
  contingency <- table(combined_df$Gender, combined_df$category)
  fisher_test_result <- fisher.test(contingency)
  
  # add stars according to significance for the P-Value
  p_value_label <- ifelse(fisher_test_result$p.value < 0.001, "***",
                        ifelse(fisher_test_result$p.value < 0.01, "**",
                               ifelse(fisher_test_result$p.value < 0.05, "*",
                                      ifelse(fisher_test_result$p.value < 0.1, "", ""))))
  
  # create a plot showing the results
  p <- ggplot(category_proportions, aes(x = Gender, y = proportion, fill = category)) +
    geom_bar(stat = "identity", position = "stack") +
    scale_fill_manual(values = c("F" = "#a6611a", "R" = "#dfc27d", "O" = "#80cdc1", "G" = "#018571"),
                      limits = c("F", "R", "O", "G"),
                      labels = c("Family", "Relationship", "Other", "Gender"))+
    theme_minimal() +
    labs(title = paste("Fisher's exact test, P-Value = ", round(fisher_test_result$p.value, 4),p_value_label, sep = ""),
         x = "Gender", y = "Share",
         fill = "Category") +
    theme(plot.caption = element_text(hjust = 0.5, size = 14),
          plot.title = element_text(hjust = 0.5),
          axis.title.x = element_text(size = 16, face = "bold"), 
          axis.title.y = element_text(size = 16, face = "bold"),
          axis.text.x = element_text(size = 16),
          axis.text.y = element_text(size = 13),
          legend.text = element_text(size = 14), 
          legend.title = element_text(size = 14))

  return(p)
}
```

### PMI results as barplots

Create plots to visualise the pmi results by applying the function "pmi_plot" to the annotated data
```{r}

#apply the function to my unmatched and matched data
deu_life <- pmi_plot(deu_life_female_cat, deu_life_male_cat)
deu_career <- pmi_plot(deu_career_female_cat, deu_career_male_cat)

matched_life <- pmi_plot(matched_life_female_cat, matched_life_male_cat)
matched_career <- pmi_plot(matched_career_female_cat, matched_career_male_cat)

# adjust axis and legend appearance for the output plots
deu_career <- deu_career + theme(legend.position = "none")
deu_life <- deu_life + theme(axis.title.y = element_blank(), axis.text.y = element_blank(), axis.ticks.y = element_blank())
matched_career <- matched_career + theme(legend.position = "none")
matched_life <- matched_life + theme(axis.title.y = element_blank(), axis.text.y = element_blank(), axis.ticks.y = element_blank())



#save the plots
ggsave(here("visualisations", "deu_life_pmi.pdf"), plot = deu_life, width = 8, height = 6, units = "in", dpi = 600)
ggsave(here("visualisations", "deu_career_pmi.pdf"), plot = deu_career, width = 8, height = 6, units = "in", dpi = 600)
ggsave(here("visualisations", "matched_life_pmi.pdf"), plot = matched_life, width = 8, height = 6, units = "in", dpi = 600)
ggsave(here("visualisations", "matched_career_pmi.pdf"), plot = matched_career, width = 8, height = 6, units = "in", dpi = 600)

#show plot output
deu_life
deu_career
matched_life
matched_career
```


### PMI results as tables

This code creates tables for the shares of categories for different subgroups
```{r}

  # add gender coloumn and Subgroup coloumn to identify the data in the table - for the life subsection data
  deu_life_female_cat$Gender <- 'female'
  deu_life_male_cat$Gender <- 'male'
  matched_life_female_cat$Gender <- 'female'
  matched_life_male_cat$Gender <- 'male'
  
  deu_life_female_cat$Subgroup <- 'whole sample, life'
  deu_life_male_cat$Subgroup <- 'whole sample, life'
  matched_life_female_cat$Subgroup <- 'matched, life'
  matched_life_male_cat$Subgroup <- 'matched, life'
  
  # add gender coloumn and Subgroup coloumn to identify the data in the table - for the career subsection data
  deu_career_female_cat$Gender <- 'female'
  deu_career_male_cat$Gender <- 'male'
  matched_career_female_cat$Gender <- 'female'
  matched_career_male_cat$Gender <- 'male'
  
  deu_career_female_cat$Subgroup <- 'whole sample, career'
  deu_career_male_cat$Subgroup <- 'whole sample, career'
  matched_career_female_cat$Subgroup <- 'matched, career'
  matched_career_male_cat$Subgroup <- 'matched, career'
  
  # combine the transformed dataframes
  combined_df <- rbind(deu_life_female_cat, deu_life_male_cat, matched_life_female_cat, matched_life_male_cat, deu_career_female_cat, deu_career_male_cat, matched_career_female_cat, matched_career_male_cat)
  
  #calculate P-Values
 p_values <- combined_df %>%
  group_by(Subgroup) %>%
  summarise(contingency = list(table(Gender, category)),
            p_value = map_dbl(contingency, ~fisher.test(.)$p.value),
            .groups = 'drop')
 
 #group and summarise data
 df_grouped <- combined_df %>%
  group_by(Subgroup, Gender, category) %>%
  summarise(count = n(), .groups = 'drop')

#Joining P-Values and Preparing the Table
df_grouped <- df_grouped %>%
  left_join(p_values %>% select(Subgroup, p_value), by = "Subgroup")

df_wide <- df_grouped %>%
  pivot_wider(names_from = category, values_from = count, values_fill = list(count = 0)) %>%
  group_by(Subgroup, Gender) %>%
  summarise(across(everything(), first), .groups = 'drop')

#Presenting the Table
df_final <- df_wide %>%
  mutate(
    `P-Value` = cell_spec(
      ifelse(p_value < 0.05, paste0("<b>", round(p_value, 4), "*</b>"), round(p_value, 4)),
      "html", escape = FALSE
    )
  )

df_final <- df_final %>%
  select(Subgroup, `P-Value`, Gender, F, R, G, O) %>%
  arrange(Subgroup, Gender) %>%
  kable("html", escape = FALSE) %>% 
  kable_styling(bootstrap_options = c("striped", "hover", "condensed")) %>%
  collapse_rows(columns = 1:2, valign = "middle")



deu_career_female_cat[, c("Gender", "Subgroup")] <- NULL
deu_career_male_cat[, c("Gender", "Subgroup")] <- NULL
deu_life_female_cat[, c("Gender", "Subgroup")] <- NULL
deu_life_male_cat[, c("Gender", "Subgroup")] <- NULL

df_final
```




Table for party subgroups, career subsection
```{r}
afd_career_female_cat$Gender <- 'female'
afd_career_male_cat$Gender <- 'male'
cdu_csu_career_female_cat$Gender <- 'female'
cdu_csu_career_male_cat$Gender <- 'male'
fdp_career_female_cat$Gender <- 'female'
fdp_career_male_cat$Gender <- 'male'
greens_career_female_cat$Gender <- 'female'
greens_career_male_cat$Gender <- 'male'
left_career_female_cat$Gender <- 'female'
left_career_male_cat$Gender <- 'male'
spd_career_female_cat$Gender <- 'female'
spd_career_male_cat$Gender <- 'male'

afd_career_female_cat$Party <- 'afd'
afd_career_male_cat$Party <- 'afd'
cdu_csu_career_female_cat$Party <- 'cdu/csu'
cdu_csu_career_male_cat$Party <- 'cdu/csu'
fdp_career_female_cat$Party <- 'fdp'
fdp_career_male_cat$Party <- 'fdp'
greens_career_female_cat$Party <- 'greens'
greens_career_male_cat$Party <- 'greens'
left_career_female_cat$Party <- 'left'
left_career_male_cat$Party <- 'left'
spd_career_female_cat$Party <- 'spd'
spd_career_male_cat$Party <- 'spd'


combined_df <- rbind(
  afd_career_female_cat, afd_career_male_cat,
  cdu_csu_career_female_cat, cdu_csu_career_male_cat,
  fdp_career_female_cat, fdp_career_male_cat,
  greens_career_female_cat, greens_career_male_cat,
  left_career_female_cat, left_career_male_cat,
  spd_career_female_cat, spd_career_male_cat
  )
expected_categories <- c("F", "G", "R", "O") 

combined_df_filled <- combined_df %>%
  group_by(Party) %>%
  complete(category = expected_categories, fill = list(Freq = 0)) %>%
  ungroup()


 p_values <- combined_df_filled %>%
  group_by(Party) %>%
  summarise(contingency = list(table(Gender, category)),
            p_value = map_dbl(contingency, ~fisher.test(.)$p.value),
            .groups = 'drop')
 
 df_grouped <- combined_df %>%
  group_by(Party, Gender, category) %>%
  summarise(count = n(), .groups = 'drop')

df_grouped <- df_grouped %>%
  left_join(p_values %>% select(Party, p_value), by = "Party")

df_wide <- df_grouped %>%
  pivot_wider(names_from = category, values_from = count, values_fill = list(count = 0)) %>%
  group_by(Party, Gender) %>%
  summarise(across(everything(), first), .groups = 'drop')%>%
  mutate(Party = factor(Party, levels = c("afd", "cdu/csu", "fdp", "spd", "greens", "left")))

df_final <- df_wide %>%
  mutate(
    `P-Value` = cell_spec(
      ifelse(p_value < .001, paste0("<b>", round(p_value, 4), "***</b>"),
             ifelse(p_value < .01, paste0("<b>", round(p_value, 4), "**</b>"),
                    ifelse(p_value < .05, paste0("<b>", round(p_value, 4), "*</b>"), 
                           ifelse(p_value < .1, paste0(round(p_value, 4), "."),
                                  paste0(round(p_value, 4), ""))))),
      "html", escape = FALSE
    )
  )


df_final <- df_final %>%
  select(Party, `P-Value`, Gender, F, R, G, O) %>%
  arrange(Party, Gender) %>%
  kable("html", escape = FALSE) %>% 
  kable_styling(bootstrap_options = c("striped", "hover", "condensed")) %>%
  collapse_rows(columns = 1:2, valign = "middle")




save_kable(df_final, file = here("visualisations", "pmi_Party_career.png"))

df_final
```

Table for party subgroups, life subsection
```{r}
afd_life_female_cat$Gender <- 'female'
afd_life_male_cat$Gender <- 'male'
cdu_csu_life_female_cat$Gender <- 'female'
cdu_csu_life_male_cat$Gender <- 'male'
fdp_life_female_cat$Gender <- 'female'
fdp_life_male_cat$Gender <- 'male'
greens_life_female_cat$Gender <- 'female'
greens_life_male_cat$Gender <- 'male'
left_life_female_cat$Gender <- 'female'
left_life_male_cat$Gender <- 'male'
spd_life_female_cat$Gender <- 'female'
spd_life_male_cat$Gender <- 'male'

afd_life_female_cat$Party <- 'afd'
afd_life_male_cat$Party <- 'afd'
cdu_csu_life_female_cat$Party <- 'cdu/csu'
cdu_csu_life_male_cat$Party <- 'cdu/csu'
fdp_life_female_cat$Party <- 'fdp'
fdp_life_male_cat$Party <- 'fdp'
greens_life_female_cat$Party <- 'greens'
greens_life_male_cat$Party <- 'greens'
left_life_female_cat$Party <- 'left'
left_life_male_cat$Party <- 'left'
spd_life_female_cat$Party <- 'spd'
spd_life_male_cat$Party <- 'spd'

combined_df <- rbind(afd_life_male_cat, afd_life_female_cat,
  cdu_csu_life_male_cat, cdu_csu_life_female_cat,
  fdp_life_male_cat, fdp_life_female_cat,
  greens_life_male_cat, greens_life_female_cat, 
  left_life_female_cat, left_life_male_cat,
  spd_life_female_cat, spd_life_male_cat)

expected_categories <- c("F", "G", "R", "O") 

combined_df_filled <- combined_df %>%
  group_by(Party) %>%
  complete(category = expected_categories, fill = list(Freq = 0)) %>%
  ungroup()


 p_values <- combined_df_filled %>%
  group_by(Party) %>%
  summarise(contingency = list(table(Gender, category)),
            p_value = map_dbl(contingency, ~fisher.test(.)$p.value),
            .groups = 'drop')
 
 df_grouped <- combined_df %>%
  group_by(Party, Gender, category) %>%
  summarise(count = n(), .groups = 'drop')

df_grouped <- df_grouped %>%
  left_join(p_values %>% select(Party, p_value), by = "Party")

df_wide <- df_grouped %>%
  pivot_wider(names_from = category, values_from = count, values_fill = list(count = 0)) %>%
  group_by(Party, Gender) %>%
  summarise(across(everything(), first), .groups = 'drop')%>%
  mutate(Party = factor(Party, levels = c("afd", "cdu/csu", "fdp", "spd", "greens", "left")))

df_final <- df_wide %>%
  mutate(
    `P-Value` = cell_spec(
      ifelse(p_value < .001, paste0("<b>", round(p_value, 4), "***</b>"),
             ifelse(p_value < .01, paste0("<b>", round(p_value, 4), "**</b>"),
                    ifelse(p_value < .05, paste0("<b>", round(p_value, 4), "*</b>"), 
                           ifelse(p_value < .1, paste0(round(p_value, 4), "."),
                                  paste0(round(p_value, 4), ""))))),
      "html", escape = FALSE
    )
  )


df_final <-df_final %>%
  select(Party, `P-Value`, Gender, F, R, G, O) %>%
  arrange(Party, Gender) %>%
  kable("html", escape = FALSE) %>%  
  kable_styling(bootstrap_options = c("striped", "hover", "condensed")) %>%
  collapse_rows(columns = 1:2, valign = "middle")




save_kable(df_final, file = here("visualisations", "pmi_Party_life.png"))

df_final
```

Table for old texts, life subsection
```{r}
matched_new_life_female_cat$Gender <- 'female'
matched_new_life_male_cat$Gender <- 'male'
matched_old_life_female_cat$Gender <- 'female'
matched_old_life_male_cat$Gender <- 'male'

matched_new_life_female_cat$Subgroup <- 'new'
matched_new_life_male_cat$Subgroup <- 'new'
matched_old_life_female_cat$Subgroup <- 'old'
matched_old_life_male_cat$Subgroup <- 'old'

combined_df <- rbind(matched_new_life_female_cat, matched_new_life_male_cat,
  matched_old_life_female_cat, matched_old_life_male_cat)



 p_values <- combined_df %>%
  group_by(Subgroup) %>%
  summarise(contingency = list(table(Gender, category)),
            p_value = map_dbl(contingency, ~fisher.test(.)$p.value),
            .groups = 'drop')
 
 df_grouped <- combined_df %>%
  group_by(Subgroup, Gender, category) %>%
  summarise(count = n(), .groups = 'drop')

df_grouped <- df_grouped %>%
  left_join(p_values %>% select(Subgroup, p_value), by = "Subgroup")

df_wide <- df_grouped %>%
  pivot_wider(names_from = category, values_from = count, values_fill = list(count = 0)) %>%
  group_by(Subgroup, Gender) %>%
  summarise(across(everything(), first), .groups = 'drop')

df_final <- df_wide %>%
  mutate(
    `P-Value` = cell_spec(
      ifelse(p_value < .001, paste0("<b>", round(p_value, 4), "***</b>"),
             ifelse(p_value < .01, paste0("<b>", round(p_value, 4), "**</b>"),
                    ifelse(p_value < .05, paste0("<b>", round(p_value, 4), "*</b>"), 
                           ifelse(p_value < .1, paste0(round(p_value, 4), "."),
                                  paste0(round(p_value, 4), ""))))),
      "html", escape = FALSE
    )
  )


df_final <- df_final %>%
  select(Subgroup, `P-Value`, Gender, F, R, G, O) %>%
  arrange(Subgroup, Gender) %>%
  kable("html", escape = FALSE) %>%  
  kable_styling(bootstrap_options = c("striped", "hover", "condensed")) %>%
  collapse_rows(columns = 1:2, valign = "middle")




save_kable(df_final, file = here("visualisations", "pmi_old_life.png"))

df_final

```

Table for old texts, career subsection
```{r}
matched_new_career_female_cat$Gender <- 'female'
matched_new_career_male_cat$Gender <- 'male'
matched_old_career_female_cat$Gender <- 'female'
matched_old_career_male_cat$Gender <- 'male'

matched_new_career_female_cat$Subgroup <- 'new'
matched_new_career_male_cat$Subgroup <- 'new'
matched_old_career_female_cat$Subgroup <- 'old'
matched_old_career_male_cat$Subgroup <- 'old'

combined_df <- rbind(matched_new_career_female_cat, matched_new_career_male_cat,
  matched_old_career_female_cat, matched_old_career_male_cat)



 p_values <- combined_df %>%
  group_by(Subgroup) %>%
  summarise(contingency = list(table(Gender, category)),
            p_value = map_dbl(contingency, ~fisher.test(.)$p.value),
            .groups = 'drop')
 
 df_grouped <- combined_df %>%
  group_by(Subgroup, Gender, category) %>%
  summarise(count = n(), .groups = 'drop')

df_grouped <- df_grouped %>%
  left_join(p_values %>% select(Subgroup, p_value), by = "Subgroup")

df_wide <- df_grouped %>%
  pivot_wider(names_from = category, values_from = count, values_fill = list(count = 0)) %>%
  group_by(Subgroup, Gender) %>%
  summarise(across(everything(), first), .groups = 'drop')

df_final <- df_wide %>%
  mutate(
    `P-Value` = cell_spec(
      ifelse(p_value < .001, paste0("<b>", round(p_value, 4), "***</b>"),
             ifelse(p_value < .01, paste0("<b>", round(p_value, 4), "**</b>"),
                    ifelse(p_value < .05, paste0("<b>", round(p_value, 4), "*</b>"), 
                           ifelse(p_value < .1, paste0(round(p_value, 4), "."),
                                  paste0(round(p_value, 4), ""))))),
      "html", escape = FALSE
    )
  )


df_final <- df_final %>%
  select(Subgroup, `P-Value`, Gender, F, R, G, O) %>%
  arrange(Subgroup, Gender) %>%
  kable("html", escape = FALSE) %>% 
  kable_styling(bootstrap_options = c("striped", "hover", "condensed")) %>%
  collapse_rows(columns = 1:2, valign = "middle")





save_kable(df_final, file = here("visualisations", "pmi_old_career.png"))

df_final
```


Table for age subgroups, life subsection
```{r}
first_cohort_life_female_cat$Gender <- 'female'
first_cohort_life_male_cat$Gender <- 'male'
second_cohort_life_female_cat$Gender <- 'female'
second_cohort_life_male_cat$Gender <- 'male'
third_cohort_life_female_cat$Gender <- 'female'
third_cohort_life_male_cat$Gender <- 'male'

first_cohort_life_female_cat$Cohort <- 'before 1949'
first_cohort_life_male_cat$Cohort <- 'before 1949'
second_cohort_life_female_cat$Cohort <- 'between 1950 and 1969'
second_cohort_life_male_cat$Cohort <- 'between 1950 and 1969'
third_cohort_life_female_cat$Cohort <- 'after 1970'
third_cohort_life_male_cat$Cohort <- 'after 1970'

combined_df <- rbind(first_cohort_life_female_cat, first_cohort_life_male_cat,
  second_cohort_life_female_cat, second_cohort_life_male_cat, 
  third_cohort_life_female_cat, third_cohort_life_male_cat)



 p_values <- combined_df %>%
  group_by(Cohort) %>%
  summarise(contingency = list(table(Gender, category)),
            p_value = map_dbl(contingency, ~fisher.test(.)$p.value),
            .groups = 'drop')
 
 df_grouped <- combined_df %>%
  group_by(Cohort, Gender, category) %>%
  summarise(count = n(), .groups = 'drop')

df_grouped <- df_grouped %>%
  left_join(p_values %>% select(Cohort, p_value), by = "Cohort")

df_wide <- df_grouped %>%
  pivot_wider(names_from = category, values_from = count, values_fill = list(count = 0)) %>%
  group_by(Cohort, Gender) %>%
  summarise(across(everything(), first), .groups = 'drop')%>%
  mutate(Cohort = factor(Cohort, levels = c("before 1949", "between 1950 and 1969", "after 1970")))


df_final <- df_wide %>%
  mutate(
    `P-Value` = cell_spec(
      ifelse(p_value < .001, paste0("<b>", round(p_value, 4), "***</b>"),
             ifelse(p_value < .01, paste0("<b>", round(p_value, 4), "**</b>"),
                    ifelse(p_value < .05, paste0("<b>", round(p_value, 4), "*</b>"), 
                           ifelse(p_value < .1, paste0(round(p_value, 4), "."),
                                  paste0(round(p_value, 4), ""))))),
      "html", escape = FALSE
    )
  )


df_final <- df_final %>%
  select(Cohort, `P-Value`, Gender, F, R, G, O) %>%
  arrange(Cohort, Gender) %>%
  kable("html", escape = FALSE) %>% 
  kable_styling(bootstrap_options = c("striped", "hover", "condensed")) %>%
  collapse_rows(columns = 1:2, valign = "middle")


save_kable(df_final, file = here("visualisations", "pmi_age_life.png"))

df_final
```


Table for age subgroups, life subsection
```{r}
first_cohort_career_female_cat$Gender <- 'female'
first_cohort_career_male_cat$Gender <- 'male'
second_cohort_career_female_cat$Gender <- 'female'
second_cohort_career_male_cat$Gender <- 'male'
third_cohort_career_female_cat$Gender <- 'female'
third_cohort_career_male_cat$Gender <- 'male'

first_cohort_career_female_cat$Cohort <- 'before 1949'
first_cohort_career_male_cat$Cohort <- 'before 1949'
second_cohort_career_female_cat$Cohort <- 'between 1950 and 1969'
second_cohort_career_male_cat$Cohort <- 'between 1950 and 1969'
third_cohort_career_female_cat$Cohort <- 'after 1970'
third_cohort_career_male_cat$Cohort <- 'after 1970'

combined_df <- rbind(first_cohort_career_female_cat, first_cohort_career_male_cat,
  second_cohort_career_female_cat, second_cohort_career_male_cat, 
  third_cohort_career_female_cat, third_cohort_career_male_cat)



 p_values <- combined_df %>%
  group_by(Cohort) %>%
  summarise(contingency = list(table(Gender, category)),
            p_value = map_dbl(contingency, ~fisher.test(.)$p.value),
            .groups = 'drop')
 
 df_grouped <- combined_df %>%
  group_by(Cohort, Gender, category) %>%
  summarise(count = n(), .groups = 'drop')

df_grouped <- df_grouped %>%
  left_join(p_values %>% select(Cohort, p_value), by = "Cohort")

df_wide <- df_grouped %>%
  pivot_wider(names_from = category, values_from = count, values_fill = list(count = 0)) %>%
  group_by(Cohort, Gender) %>%
  summarise(across(everything(), first), .groups = 'drop')%>%
  mutate(Cohort = factor(Cohort, levels = c("before 1949", "between 1950 and 1969", "after 1970")))


df_final <- df_wide %>%
  mutate(
    `P-Value` = cell_spec(
      ifelse(p_value < .001, paste0("<b>", round(p_value, 4), "***</b>"),
             ifelse(p_value < .01, paste0("<b>", round(p_value, 4), "**</b>"),
                    ifelse(p_value < .05, paste0("<b>", round(p_value, 4), "*</b>"), 
                           ifelse(p_value < .1, paste0(round(p_value, 4), "."),
                                  paste0(round(p_value, 4), ""))))),
      "html", escape = FALSE
    )
  )


df_final <- df_final %>%
  select(Cohort, `P-Value`, Gender, F, R, G, O) %>%
  arrange(Cohort, Gender) %>%
  kable("html", escape = FALSE) %>% 
  kable_styling(bootstrap_options = c("striped", "hover", "condensed")) %>%
  collapse_rows(columns = 1:2, valign = "middle")



save_kable(df_final, file = here("visualisations", "pmi_age_career.png"))

df_final

```

table for sensitivity analysis, career section
```{r}
S_deu_career_female_cat$Gender <- 'female'
S_deu_career_male_cat$Gender <- 'male'
S2_deu_career_female_cat$Gender <- 'female'
S2_deu_career_male_cat$Gender <- 'male'
S3_deu_career_female_cat$Gender <- 'female'
S3_deu_career_male_cat$Gender <- 'male'
S4_deu_career_female_cat$Gender <- 'female'
S4_deu_career_male_cat$Gender <- 'male'
deu_career_female_cat$Gender <- 'female'
deu_career_male_cat$Gender <- 'male'

S_deu_career_female_cat$Threshold <- '0.001'
S_deu_career_male_cat$Threshold <- '0.001'
S2_deu_career_female_cat$Threshold <- '0.0007'
S2_deu_career_male_cat$Threshold <- '0.0007'
S3_deu_career_female_cat$Threshold <- '0.0004'
S3_deu_career_male_cat$Threshold <- "0.0004"
S4_deu_career_female_cat$Threshold <- '0.0003'
S4_deu_career_male_cat$Threshold <- "0.0003"
deu_career_female_cat$Threshold <- '0.0002'
deu_career_male_cat$Threshold <- '0.0002'

combined_df <- rbind(deu_career_female_cat, deu_career_male_cat, S_deu_career_female_cat, S_deu_career_male_cat,
  S2_deu_career_female_cat, S2_deu_career_male_cat, 
  S3_deu_career_female_cat, S3_deu_career_male_cat)





 p_values <- combined_df %>%
  group_by(Threshold) %>%
  summarise(contingency = list(table(Gender, category)),
            p_value = map_dbl(contingency, ~fisher.test(.)$p.value),
            .groups = 'drop')
 
 df_grouped <- combined_df %>%
  group_by(Threshold, Gender, category) %>%
  summarise(count = n(), .groups = 'drop')

df_grouped <- df_grouped %>%
  left_join(p_values %>% select(Threshold, p_value), by = "Threshold")

df_wide <- df_grouped %>%
  pivot_wider(names_from = category, values_from = count, values_fill = list(count = 0)) %>%
  group_by(Threshold, Gender) %>%
  summarise(
    across(everything(), first),
    "Number of Words" = sum(F, R, G, O),  # Assuming F, R, G, O are your category columns containing counts
    .groups = 'drop'
  )

df_final <- df_wide %>%
  mutate(
    `P-Value` = cell_spec(
      ifelse(p_value < .001, paste0("<b>", round(p_value, 4), "***</b>"),
             ifelse(p_value < .01, paste0("<b>", round(p_value, 4), "**</b>"),
                    ifelse(p_value < .05, paste0("<b>", round(p_value, 4), "*</b>"), 
                           ifelse(p_value < .1, paste0(round(p_value, 4), "."),
                                  paste0(round(p_value, 4), ""))))),
      "html", escape = FALSE
    )
  )


df_final <- df_wide %>%
  mutate(
    `P-Value` = cell_spec(
      ifelse(p_value < .001, paste0("<b>", round(p_value, 4), "***</b>"),
             ifelse(p_value < .01, paste0("<b>", round(p_value, 4), "**</b>"),
                    ifelse(p_value < .05, paste0("<b>", round(p_value, 4), "*</b>"),
                           ifelse(p_value < .1, paste0(round(p_value, 4), "."),
                                  paste0(round(p_value, 4), ""))))),
      "html", escape = FALSE
    )
  ) %>%
  select(Threshold, `P-Value`, Gender, F, R, G, O, "Number of Words") %>%
  arrange(Threshold, Gender) %>%
  kable("html", escape = FALSE) %>%
  kable_styling(bootstrap_options = c("striped", "hover", "condensed")) %>%
  collapse_rows(columns = 1:2, valign = "middle")


kableExtra::save_kable(df_final, file = here("visualisations", "sensitivity_career.png"))

df_final

```


table for sensitivity analysis, career section
```{r}
S_deu_life_female_cat$Gender <- 'female'
S_deu_life_male_cat$Gender <- 'male'
S2_deu_life_female_cat$Gender <- 'female'
S2_deu_life_male_cat$Gender <- 'male'
S3_deu_life_female_cat$Gender <- 'female'
S3_deu_life_male_cat$Gender <- 'male'
S4_deu_life_female_cat$Gender <- 'female'
S4_deu_life_male_cat$Gender <- 'male'
deu_life_female_cat$Gender <- 'female'
deu_life_male_cat$Gender <- 'male'

S_deu_life_female_cat$Threshold <- '0.001'
S_deu_life_male_cat$Threshold <- '0.001'
S2_deu_life_female_cat$Threshold <- '0.0007'
S2_deu_life_male_cat$Threshold <- '0.0007'
S3_deu_life_female_cat$Threshold <- '0.0004'
S3_deu_life_male_cat$Threshold <- "0.0004"
S4_deu_life_female_cat$Threshold <- '0.0003'
S4_deu_life_male_cat$Threshold <- "0.0003"
deu_life_female_cat$Threshold <- '0.0002'
deu_life_male_cat$Threshold <- '0.0002'

combined_df <- rbind(deu_life_female_cat, deu_life_male_cat, S_deu_life_female_cat, S_deu_life_male_cat,
  S2_deu_life_female_cat, S2_deu_life_male_cat, 
  S3_deu_life_female_cat, S3_deu_life_male_cat)



 p_values <- combined_df %>%
  group_by(Threshold) %>%
  summarise(contingency = list(table(Gender, category)),
            p_value = map_dbl(contingency, ~fisher.test(.)$p.value),
            .groups = 'drop')
 
 df_grouped <- combined_df %>%
  group_by(Threshold, Gender, category) %>%
  summarise(count = n(), .groups = 'drop')

df_grouped <- df_grouped %>%
  left_join(p_values %>% select(Threshold, p_value), by = "Threshold")

df_wide <- df_grouped %>%
  pivot_wider(names_from = category, values_from = count, values_fill = list(count = 0)) %>%
  group_by(Threshold, Gender) %>%
  summarise(
    across(everything(), first),
    "Number of Words" = sum(F, R, G, O),  # Assuming F, R, G, O are your category columns containing counts
    .groups = 'drop'
  )

df_final <- df_wide %>%
  mutate(
    `P-Value` = cell_spec(
      ifelse(p_value < .001, paste0("<b>", round(p_value, 4), "***</b>"),
             ifelse(p_value < .01, paste0("<b>", round(p_value, 4), "**</b>"),
                    ifelse(p_value < .05, paste0("<b>", round(p_value, 4), "*</b>"), 
                           ifelse(p_value < .1, paste0(round(p_value, 4), "."),
                                  paste0(round(p_value, 4), ""))))),
      "html", escape = FALSE
    )
  )


df_final <- df_wide %>%
  mutate(
    `P-Value` = cell_spec(
      ifelse(p_value < .001, paste0("<b>", round(p_value, 4), "***</b>"),
             ifelse(p_value < .01, paste0("<b>", round(p_value, 4), "**</b>"),
                    ifelse(p_value < .05, paste0("<b>", round(p_value, 4), "*</b>"),
                           ifelse(p_value < .1, paste0(round(p_value, 4), "."),
                                  paste0(round(p_value, 4), ""))))),
      "html", escape = FALSE
    )
  ) %>%
  select(Threshold, `P-Value`, Gender, F, R, G, O, "Number of Words") %>%
  arrange(Threshold, Gender) %>%
  kable("html", escape = FALSE) %>%
  kable_styling(bootstrap_options = c("striped", "hover", "condensed")) %>%
  collapse_rows(columns = 1:2, valign = "middle")


kableExtra::save_kable(df_final, file = here("visualisations", "sensitivity_life.png"))

df_final

```
